requirements.txt
cv2
matplotlib
numpy
Pillow
torch
torchvision
tensorboard
future
pandas
tqdm
opencv
imageio
sklearn